+++ example/lib/main.dart
@@ -1,5 +1,4 @@
 import 'package:dio_request_inspector/dio_request_inspector.dart';
-import 'package:dio_request_inspector/presentation/main/page/main_page.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'dio_request_inspector_setting.dart';